Move inspect into tools
