--- internal/inspect/go.mod
@@ -1,39 +1,0 @@
-module go.opentelemetry.io/auto/internal/inspect
-
-go 1.17
-
-require (
-	github.com/docker/docker v24.0.5+incompatible
-	github.com/go-logr/logr v1.2.4
-	github.com/go-logr/stdr v1.2.2
-	github.com/hashicorp/go-version v1.6.0
-	go.opentelemetry.io/auto v0.0.0-00010101000000-000000000000
-	golang.org/x/sync v0.3.0
-)
-
-require (
-	github.com/Microsoft/go-winio v0.6.1 // indirect
-	github.com/cilium/ebpf v0.11.0 // indirect
-	github.com/docker/distribution v2.8.2+incompatible // indirect
-	github.com/docker/go-connections v0.4.0 // indirect
-	github.com/docker/go-units v0.5.0 // indirect
-	github.com/go-logr/zapr v1.2.4 // indirect
-	github.com/gogo/protobuf v1.3.2 // indirect
-	github.com/moby/term v0.5.0 // indirect
-	github.com/morikuni/aec v1.0.0 // indirect
-	github.com/opencontainers/go-digest v1.0.0 // indirect
-	github.com/opencontainers/image-spec v1.0.2 // indirect
-	github.com/pkg/errors v0.9.1 // indirect
-	go.uber.org/multierr v1.10.0 // indirect
-	go.uber.org/zap v1.25.0 // indirect
-	golang.org/x/arch v0.5.0 // indirect
-	golang.org/x/exp v0.0.0-20230224173230-c95f2b4c22f2 // indirect
-	golang.org/x/mod v0.8.0 // indirect
-	golang.org/x/net v0.15.0 // indirect
-	golang.org/x/sys v0.12.0 // indirect
-	golang.org/x/time v0.3.0 // indirect
-	golang.org/x/tools v0.6.0 // indirect
-	gotest.tools/v3 v3.5.0 // indirect
-)
-
-replace go.opentelemetry.io/auto => ../../
