+++ internal/tools/inspect/app.go
