+++ internal/tools/inspect/builder.go
