+++ internal/tools/inspect/cache.go
