+++ internal/tools/inspect/cmd/offsetgen/verions.go
