+++ internal/tools/inspect/inspector.go
