+++ internal/tools/inspect/offsets.go
