+++ internal/tools/inspect/render.go
