+++ internal/tools/inspect/templates/google.golang.org/grpc/go.mod.tmpl
