+++ internal/tools/inspect/templates/google.golang.org/grpc/main.go.tmpl
