+++ internal/tools/inspect/templates/net/http/go.mod.tmpl
