+++ internal/tools/inspect/templates/net/http/main.go.tmpl
