+++ internal/tools/inspect/templates/runtime/go.mod.tmpl
