+++ internal/tools/inspect/templates/runtime/main.go.tmpl
